+++ controllers/rcj_soccer_ball/rcj_soccer_ball.py
@@ -1,12 +1,9 @@
-import struct
-
 from controller import Robot
 
 robot = Robot()
 ball_emitter = robot.getDevice("ball emitter")
 
-data = [True]  # Packet cannot be empty
-packet = struct.pack("?", *data)
+data = "x"  # Packet cannot be empty
 
 while robot.step(32) != -1:
-    ball_emitter.send(packet)
+    ball_emitter.send(data)